organize supabase projects lib
