--- src/lib/sepana/api/index.ts
@@ -1,2 +1,0 @@
-export * from './generic'
-export * from './projects'
