+++ src/models/supabaseProject.ts
@@ -52,20 +52,3 @@ export type SBProject = {
     'description' | 'logoUri' | 'name' | 'tags' | 'archived'
   >
 >
-
-export type SepanaQueryResponse<T> = {
-  hits: {
-    total: {
-      value: number
-      relation: string
-    }
-    max_score: number
-    hits: {
-      _index: string
-      _type: string
-      _id: string
-      _score: number
-      _source: T
-    }[]
-  }
-}
