+++ src/pages/api/projects/update.page.ts
@@ -1,7 +1,7 @@
-import { updateSBProjects } from 'lib/sepana/update'
+import { updateSBProjects } from 'lib/api/supabase/projects'
 import { NextApiHandler } from 'next'
 
-// Synchronizes the Sepana engine with the latest Juicebox Subgraph/IPFS data
+// Synchronizes the Supabase db with the latest Juicebox Subgraph/IPFS data
 const handler: NextApiHandler = async (_, res) => {
   await updateSBProjects(res, false)
 }
